+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Upcoming]
+### Changed
+* Config field "Estimation/Merge/barcodes_file" now accepts paths, relative to config directory, as well as path with "~/"
+### Added
+* To improve reproducibility, config file is now copied to log directory
+
 ## [0.8.1] - 2018-04-06
 ### Changed
 * Fixed some bugs in dropestr